feat: introduce a mock proof mode
