+++ crates/proposer-client/src/network_prover.rs
@@ -1,27 +1,10 @@
 use std::time::Duration;
 
 use alloy_primitives::B256;
-use anyhow::{Context, Error};
+use anyhow::Context;
 use sp1_sdk::{
     NetworkProver, Prover, SP1ProofWithPublicValues, SP1VerificationError, SP1VerifyingKey,
 };
-
-/// This prover waits for the SP1 cluster generated
-/// AggregationProof based on the proof id.
-#[tonic::async_trait]
-pub trait AggregationProver {
-    async fn wait_for_proof(
-        &self,
-        request_id: B256,
-        timeout: Option<Duration>,
-    ) -> Result<SP1ProofWithPublicValues, anyhow::Error>;
-
-    fn verify_aggregated_proof(
-        &self,
-        proof: &SP1ProofWithPublicValues,
-        vkey: &SP1VerifyingKey,
-    ) -> Result<(), SP1VerificationError>;
-}
 
 #[tonic::async_trait]
 impl AggregationProver for NetworkProver {
@@ -29,7 +12,7 @@ impl AggregationProver for NetworkProver {
         &self,
         request_id: B256,
         timeout: Option<Duration>,
-    ) -> Result<SP1ProofWithPublicValues, Error> {
+    ) -> anyhow::Result<SP1ProofWithPublicValues> {
         self.wait_proof(request_id, timeout).await
     }
